more stuff (read below)
1. fix job comparator for max heap. the job with the _lowest_ priority
should be first
2. fix add/remove waiting job functions
3. add_job takes a tcb* now
4. get_next_job returns a tcb* now
5. move out tests from jobsim.c to test.c
6. rename jobsim to benchmark
