+++ asst1/data_structure.h
@@ -70,7 +70,7 @@ void hash_insert(hash_table * h, void * t, int id);
 void * hash_find(hash_table * h, int id);
 void hash_delete(hash_table *h, int id);
 
-// Heap
+// Max Heap
 
 m_heap *m_heap_init(cmp_fn cmp);
 int m_heap_is_empty(m_heap *h);
