+++ asst1/scheduler.h
@@ -36,11 +36,11 @@ typedef struct multiLevelQueue {
 
 multi_queue * m_queue_init(int num_levels, int time_delta, int base_time);
 void init_job(void * element, multi_queue * m_q);
-void * get_next_job(multi_queue * m_q);
+tcb * get_next_job(multi_queue * m_q);
 int is_empty_m_queue(multi_queue * m_q);
 int get_interval_time(int level, multi_queue * m_q);
 void cleanup_m_queue(multi_queue * m_q);
-void add_job(void * element, multi_queue * m_q);
+void add_job(tcb * element, multi_queue * m_q);
 
 // Scheduler
 
@@ -60,5 +60,6 @@ typedef struct scheduler {
 
 sched *sched_init(int num_queue_levels, int alarm_time_delta,
     int alarm_base_time);
-void add_waiting_job(sched* s, tcb *t, hash_table *jobs, int id);
+void add_waiting_job(tcb *t, hash_table *jobs, int id);
+tcb *remove_waiting_job(hash_table *jobs, int id);
